+++ plugins/destination/postgresql/client/deletestale_test.go
@@ -12,7 +12,7 @@ import (
 func TestDeleteStale(t *testing.T) {
 	ctx := context.Background()
 	client, err := New(ctx, getTestLogger(t), specs.Destination{
-		WriteMode: specs.WriteModeOverwriteDeletestale,
+		WriteMode: specs.WriteModeOverwriteDeleteStale,
 		Spec: &Spec{
 			ConnectionString: getTestConnection(),
 		},
@@ -51,7 +51,7 @@ func TestDeleteStale(t *testing.T) {
 	// TODO: figure normal pgx<->go time normal conversion
 	syncTime := time.Now().UTC().Add(-24*30*12 * time.Hour)
 	// syncTime = syncTime.Add(time.Second * 2)
-	if err := c.DeleteStale(ctx, []string{"simple_table"}, testData[schema.CqSourceName.Name].(string), syncTime); err != nil {
+	if err := c.DeleteStale(ctx, "simple_table", testData[schema.CqSourceName.Name].(string), syncTime); err != nil {
 		t.Fatalf("failed to delete stale data: %v", err)
 	}
 	totalResults, err = selectTableAsJson(ctx, c.conn, testTable.Name, &results)
@@ -65,7 +65,7 @@ func TestDeleteStale(t *testing.T) {
 		t.Fatalf("got %d, expected json_agg to return list with 1 entries", len(results))
 	}
 	syncTime = time.Now().UTC()
-	if err := c.DeleteStale(ctx, []string{"simple_table"}, testData[schema.CqSourceName.Name].(string), syncTime); err != nil {
+	if err := c.DeleteStale(ctx, "simple_table", testData[schema.CqSourceName.Name].(string), syncTime); err != nil {
 		t.Fatalf("failed to delete stale data: %v", err)
 	}
 	totalResults, err = selectTableAsJson(ctx, c.conn, testTable.Name, &results)
