+++ plugins/destination/postgresql/client/drop_test.go
@@ -11,7 +11,7 @@ import (
 func TestDrop(t *testing.T) {
 	ctx := context.Background()
 	client, err := New(ctx, getTestLogger(t), specs.Destination{
-		WriteMode: specs.WriteModeOverwriteDeletestale,
+		WriteMode: specs.WriteModeOverwriteDeleteStale,
 		Spec: &Spec{
 			ConnectionString: getTestConnection(),
 		},
