+++ plugins/destination/postgresql/client/migrate_test.go
@@ -12,7 +12,7 @@ import (
 func TestMigrate(t *testing.T) {
 	ctx := context.Background()
 	client, err := New(ctx, getTestLogger(t), specs.Destination{
-		WriteMode: specs.WriteModeOverwriteDeletestale,
+		WriteMode: specs.WriteModeOverwriteDeleteStale,
 		Spec: &Spec{
 			ConnectionString: getTestConnection(),
 		},
