+++ plugins/destination/test/main.go
@@ -1,10 +1,15 @@
 package main
 
 import (
-	"github.com/cloudquery/cloudquery/plugins/destination/test/plugin"
+	"github.com/cloudquery/cloudquery/plugins/destination/test/client"
+	"github.com/cloudquery/plugin-sdk/plugins"
 	"github.com/cloudquery/plugin-sdk/serve"
 )
 
+var (
+	Version = "Development"
+)
+
 func main() {
-	serve.Destination(plugin.New())
+	serve.Destination(plugins.NewDestinationPlugin("test", Version, client.New))
 }
