+++ plugins/source/test/go.mod
@@ -3,7 +3,7 @@ module github.com/cloudquery/cloudquery/plugins/source/test
 go 1.19
 
 require (
-	github.com/cloudquery/plugin-sdk v0.7.2
+	github.com/cloudquery/plugin-sdk v0.12.0
 	github.com/rs/zerolog v1.28.0
 )
